requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
absl-py==1.2.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
cycler==0.11.0
flatbuffers==1.12
fonttools==4.34.4
gast==0.4.0
google-auth-oauthlib==0.4.6
google-auth==2.9.1
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.47.0
h5py==3.7.0
idna==3.3
imageio==2.21.0
importlib-metadata==4.12.0
joblib==1.1.0
keras-preprocessing==1.1.2
keras==2.9.0
kiwisolver==1.4.4
libclang==14.0.6
markdown==3.4.1
markupsafe==2.1.1
matplotlib==3.5.2
networkx==2.8.5
numpy==1.23.1
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==23.2; python_version >= '3.7'
pandas==2.2.0
pillow==9.2.0
protobuf==3.19.4
pyasn1-modules==0.2.8
pyasn1==0.4.8
pydot==1.4.2
pydotplus==2.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.4
pywavelets==1.3.0
requests-oauthlib==1.3.1
requests==2.28.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.9.0
setuptools==69.0.3; python_version >= '3.8'
six==1.16.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboard==2.9.1
tensorflow-addons==0.17.1
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.1
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2022.8.3
typeguard==2.13.3
typing-extensions==4.3.0
tzdata==2023.4; python_version >= '2'
urllib3==1.26.11
werkzeug==2.2.1
wheel==0.42.0; python_version >= '3.7'
wrapt==1.14.1
zipp==3.8.1
